Powershell script that mimics ransomware behavior
Powershell script to encrypt files and another to decrypt the files
using a public/private key pair. The purpose of the scripts are to
safely test your network defense against a ransomware attack. The
encrypt script, attacks network shares, but only drives that you have
mapped, starting with the lowest drive letter first. This gives you
control over what files are attacked. It also makes a copy of every file
before it encrypts them. Because it uses file stream, it doesn't delete
the original file, it simply overwrites it. The decrypt script decrypts
the files so you can undo any damage that may have been caused.